refactor: dedupe batch header / blob helpers, single-source on common/batch (#945)
